refactor: rename local xcm_config to xcm_barrier
